Threading algorithm with separate event, collection and send queues
Implemented an algorithm where each matric to be collected is added as
an event to a priority event queue. The priority event queue is sorted
based on the next process time of the matric and is watched by a single
thread. Whenever a matric is to be collected, it is moved to collection
queue, where a collection thread pool is used to wait for collection on
I/O. Once the collection is done, the matric is added back as an event
to priority event queue. Conditionally, if the number od samples
collected is same as aggregation_size specified by user, the matric is
also added to send queue.