Adding post-install-setup.sh so there's a single command for setup
This is just a helper for folks who are doing development and
*IS NOT* intended as a replacement for proper packaging (no pip
is not considered proper packaging).  This is intended to fail if
anything is unexpected for any reason, and assumptions can be
overridden if needed for some reason.

This file should not be included, or shipped, with proper packages

Signed-off-by: John 'Warthog9' Hawley <“jhawley@vmware.com”>